+++ backend/mail-server/src/main/java/news/bombomemail/article/domain/Article.java
@@ -1,4 +1,4 @@
-package news.bombomemail.article;
+package news.bombomemail.article.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
