+++ backend/mail-server/src/main/java/news/bombomemail/article/repository/ArticleRepository.java
@@ -1,5 +1,6 @@
-package news.bombomemail.article;
+package news.bombomemail.article.repository;
 
+import news.bombomemail.article.domain.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
