+++ backend/mail-server/src/main/java/news/bombomemail/email/service/EmailService.java
@@ -10,7 +10,7 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import news.bombomemail.article.ArticleService;
+import news.bombomemail.article.service.ArticleService;
 import news.bombomemail.email.util.EmailContentExtractor;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
