+++ backend/mail-server/src/main/java/news/bombomemail/member/domain/Member.java
@@ -1,4 +1,4 @@
-package news.bombomemail.member;
+package news.bombomemail.member.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -16,6 +16,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import news.bombomemail.common.BaseEntity;
+import news.bombomemail.member.domain.Gender;
 
 @Entity
 @Getter
