+++ backend/mail-server/src/main/java/news/bombomemail/member/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package news.bombomemail.member;
+package news.bombomemail.member.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
