+++ backend/mail-server/src/main/java/news/bombomemail/newsletter/domain/Newsletter.java
@@ -1,4 +1,4 @@
-package news.bombomemail.newsletter;
+package news.bombomemail.newsletter.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
