+++ backend/mail-server/src/main/java/news/bombomemail/newsletter/repository/NewsletterRepository.java
@@ -1,6 +1,7 @@
-package news.bombomemail.newsletter;
+package news.bombomemail.newsletter.repository;
 
 import java.util.Optional;
+import news.bombomemail.newsletter.domain.Newsletter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NewsletterRepository extends JpaRepository<Newsletter, Long> {
