+++ backend/mail-server/src/test/java/news/bombomemail/article/ArticleServiceTest.java
@@ -7,12 +7,15 @@ import jakarta.mail.internet.InternetAddress;
 import jakarta.mail.internet.MimeMessage;
 import java.util.List;
 import java.util.Properties;
+import news.bombomemail.article.domain.Article;
+import news.bombomemail.article.repository.ArticleRepository;
+import news.bombomemail.article.service.ArticleService;
 import news.bombomemail.email.util.EmailContentExtractor;
-import news.bombomemail.member.Gender;
-import news.bombomemail.member.Member;
-import news.bombomemail.member.MemberRepository;
-import news.bombomemail.newsletter.Newsletter;
-import news.bombomemail.newsletter.NewsletterRepository;
+import news.bombomemail.member.domain.Gender;
+import news.bombomemail.member.domain.Member;
+import news.bombomemail.member.repository.MemberRepository;
+import news.bombomemail.newsletter.domain.Newsletter;
+import news.bombomemail.newsletter.repository.NewsletterRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
