+++ backend/mail-server/src/test/java/news/bombomemail/email/service/EmailServiceTest.java
@@ -4,15 +4,15 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.List;
-import news.bombomemail.article.Article;
-import news.bombomemail.article.ArticleRepository;
-import news.bombomemail.article.ArticleService;
+import news.bombomemail.article.domain.Article;
+import news.bombomemail.article.repository.ArticleRepository;
+import news.bombomemail.article.service.ArticleService;
 import news.bombomemail.email.EmailConfig;
-import news.bombomemail.member.Member;
-import news.bombomemail.member.MemberRepository;
-import news.bombomemail.member.Gender;
-import news.bombomemail.newsletter.Newsletter;
-import news.bombomemail.newsletter.NewsletterRepository;
+import news.bombomemail.member.domain.Member;
+import news.bombomemail.member.repository.MemberRepository;
+import news.bombomemail.member.domain.Gender;
+import news.bombomemail.newsletter.domain.Newsletter;
+import news.bombomemail.newsletter.repository.NewsletterRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
